fix(gui): override settings by cmd options
